Extract wgpu dispatch helpers into bitnet-dispatch-core
