+++ Cargo.toml
@@ -100,6 +100,7 @@ members = [
   "crates/bitnet-vulkan",        # Vulkan GLSL compute shaders for GPU inference
   "crates/bitnet-vulkan-shaders",# SRP: embedded Vulkan GLSL shader source catalog
   "crates/bitnet-wgpu",          # WebGPU compute backend for cross-platform GPU inference
+  "crates/bitnet-dispatch-core", # SRP: backend-agnostic compute dispatch sizing/contracts
   "crates/bitnet-bench-receipts", # SRP: benchmark receipt model + JSONL persistence
   "crates/bitnet-wgpu-bench",    # Benchmarking and profiling harness for wgpu compute kernels
   "crates/bitnet-wgpu-runner",   # WebGPU kernel runner for GPU compute dispatch
