+++ crates/bitnet-dispatch-core/Cargo.toml
@@ -1,0 +1,12 @@
+[package]
+name = "bitnet-dispatch-core"
+version = "0.1.0"
+edition = "2024"
+rust-version = "1.92.0"
+description = "Backend-agnostic compute dispatch sizing and recording primitives"
+license = "MIT OR Apache-2.0"
+
+[dependencies]
+
+[features]
+default = []
